git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-SVN/ARQ/trunk@1117287 13f79535-47bb-0310-9956-ffa450edef68

+++ src/com/hp/hpl/jena/sparql/expr/nodevalue/Functions.java
@@ -382,7 +382,7 @@ public class Functions
             // No decimal round in Java 1.4
             return (int)Math.round(nv.getDecimal().doubleValue()) ;
         
-        if ( nv.isFloat()   ) return (int)Math.round(nv.getFloat()) ;
+        if ( nv.isFloat()   ) return Math.round(nv.getFloat()) ;
         if ( nv.isDouble()  ) return (int)Math.round(nv.getDouble()) ;
         throw new ExprEvalException("Not a number:"+nv) ;
     }
